Add testcase

+++ tddkata/Cylinder.cs
@@ -5,9 +5,9 @@ namespace tddkata
     public class Cylinder : IShape
     {
         private double radius;
-        private int length;
+        private double length;
 
-        public Cylinder(double radius, int length)
+        public Cylinder(double radius, double length)
         {
             this.radius = radius;
             this.length = length;
